Composing with Components
Composition is the ability to encapsulate a bunch of elements in a component and then easily reuse them and then also the ability to configure those components differently and independently via props() are two really important and fundamental property to understand composition of models in React.
Using Components encourages us to build our app using composition instead of inheritance.