Bind form input field to a component state
In this video, we will bind our input field to to whatever the value of a certain property of a state is. This will allows us to update the UI based on the form data. So the input value will be whatever the query property state is
So we will have our input field state living inside of a component, so this allow us to update our Ui based on that state.

A good rule of the thumb is if you want the form data to update the UI in any other way besides just updating the input field itself, you should make your form a controlled component where react is controlling the state of the input field. Otherwise you can go ahead and put the data in the Dom somewhere, then you can grab the data when you need it

To recap how user input affects the ListContacts component's own state:

The user enters text into the input field.
The onChange event listener invokes the updateQuery() function.
updateQuery() then calls setState(), merging in the new state to update the component's internal state.
Because its state has changed, the ListContacts component re-renders.