Update state in Controlled component
We want to update the UI(filter our list) based on what the user type in the form input field
So in the previous commit, we bind the input field to the component query property state, and whenever the component state changes(onChange input method triggers an arrow function that calls a method that updates the state using setState), it causes a re-render and then we go ahead and filter our contacts based on whatever the query property value is

They are called Controlled components is because React is controlling the state of the form.
Although implementing controlled components requires more typing of code but they do have some benefits like supporting instant input validation for our form, enforce input formats
All these benefits have to do with updating the UI based on some user input which is the heart of not only controlled components but react in general. If the state of our application changes, then our UI updates based of that new state.