componentDidMount Lifecycle Event
componentDidMount() is the lifecycle hook that is run right after the component is added to the DOM and should ONLY be used if you're fetching remote data or doing an Ajax request

You shouldn't make Ajax requests in the render method. This is because you don't have complete control over when the render() method will be invoked. The render method needs to be free of side effects. It should never fetch ajax requests or do anything that is asynchronous in nature. It should only receive props and return the description of the UI.

SIDE NOTE FOR CODE AFTER THIS COMMIT:
- Now we're syncing our contacts back to the server, any delete commands will remove all of them. Since we aren't able to add new ones yet, you'll be kinda stuck if you delete all of them.
Before, when we delete a contact, the UI re-rendrs and show the updated state but when we reload the page, the state goes back to the default list that we had intialized.
- But now, we have it fetch from database, when we delete a contact, we also delete from server. so when we refresh, the componentDidMount event, we fetch the update list from server. Now we're syncing our contacts back to the server
- FYI, we did not have our server connect to a realife database. So you can still get your default list from the server, when you restart the back-end server