Dynamically Render Pages - 2
From the previous commit, the user is not able to modify the screen state, so we add a button for them to do so in this commit

However, when we click the browser back button, we did not go back to the previous state where we show the list of contacts. We broke the back button of our app.
This is not good because this button is an essential for the web
When the user clicks the back button, they expect to see the previous screen. It gets tricky to do this with components state all on our own.
This is why react Router exitsts so that we can keep our UI and URL in sync
It keeps all expectations around Link Screens and URLs intact