The Link Component
For this commit, we change the button we added for the commit(Dynamically Render Pages - 2) to a Link Component. So now instead of clicking a <a> tag we click on the Link Component.
So now, we have a BrowserRouter and our Link Component and when we click on the Link Component, the Link Component talks to the BrowserRouter and tells it to update the URL. so we will have http://localhost:3000/create and the back button will work as the user expects instead of us breaking the button
So not we do not need to add a call callback we passed as prop for the onclick in the previous time where we used <a> HTML tag

NOTE: You'll use it in place of anchor tags (a) as you're typically used to. React Router's <Link> component is a great way to make navigation through your app accessible for users.

You can read more information about Link in the official docs.
https://reactrouter.com/web/api/Link
https://github.com/ReactTraining/react-router/blob/master/packages/react-router-dom/modules/Link.js