The Route Component
Route component takes a path prop that will match the browser URL or not. Components wrapped in the Router component will only render the Component UI when it matches (at least some initial part of) the URL. If the 'exact' flag is set, the path will only match when it exactly matches the URL. It does not render anything it does not match. Just like the code (in commit 'Dynamically Render Pages - 1') that we have checking the component state, to decide which screen we wanted to render, Route does thesame type of thing. But instead of checking for component's state, it checks for the URL. This means that the browser back button will still work.

In summary, the Route component is a critical piece of building an application with React Router because it's the component which is going to decide which components are rendered based on the current URL path.