Finishing The Contact Form - 1
In this commit, we just build out the UI for the form with the css style we want. We have not make React control the state of this form just yet neither have we validated the input. One thing to note here is the ImageInput Component

The ImageInput component is a custom <input> that dynamically reads and resizes image files before submitting them to the server as data URLs. It also shows a preview of the image. We chose to give this component to you rather than build it ourselves because it contains features related to files and images on the web that aren't crucial to your education in this context. If you're curious, feel free to dive into the code, but know it's not a requirement.